Move TestableVisitor to separate file to allow reuse
