+++ fluent.syntax/src/test/kotlin/org/projectfluent/syntax/visitor/VisitorTest.kt
@@ -2,9 +2,6 @@ package org.projectfluent.syntax.visitor
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import org.projectfluent.syntax.ast.Pattern
-import org.projectfluent.syntax.ast.TextElement
-import org.projectfluent.syntax.ast.Variant
 import org.projectfluent.syntax.parser.FluentParser
 
 class VisitorTest {
@@ -26,29 +23,5 @@ class VisitorTest {
         assertEquals(3, visitor.wordCount)
         assertEquals(2, visitor.patternCount)
         assertEquals(1, visitor.variantCount)
-    }
-}
-
-private class TestableVisitor : Visitor() {
-    var patternCount = 0
-    var variantCount = 0
-    var wordCount = 0
-
-    fun visitPattern(node: Pattern) {
-        patternCount++
-        genericVisit(node)
-    }
-
-    fun visitVariant(node: Variant) {
-        variantCount++
-        genericVisit(node)
-    }
-
-    fun visitTextElement(node: TextElement) {
-        wordCount += WORDS.findAll(node.value).count()
-    }
-
-    private companion object {
-        private val WORDS = Regex("\\w+")
     }
 }
